Pythonサンプルコード: 数値型の縮小変換により実行時例外が起きる問題の修正（32bit環境） (#35)
* use int32 for numpy.repeat repeats

* add comment

+++ example/python/forwarder.py
@@ -146,11 +146,12 @@ class Forwarder:
             if p.phoneme in unvoiced_mora_phoneme_list:
                 f0_list[i] = 0
 
+        # use numpy.int32 as the number of repeats to avoid casting int64 to int32 in numpy internal
         phoneme = numpy.repeat(
-            phoneme_list_s, numpy.round(phoneme_length * rate).astype(numpy.int64)
+            phoneme_list_s, numpy.round(phoneme_length * rate).astype(numpy.int32)
         )
         f0 = numpy.repeat(
-            f0_list, numpy.round(phoneme_length_sa * rate).astype(numpy.int64)
+            f0_list, numpy.round(phoneme_length_sa * rate).astype(numpy.int32)
         )
 
         # forward decode
